move new functions to separate file + docstring + fix copyright years
