+++ planning/behavior_velocity_dynamic_obstacle_stop_module/src/collision.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 TIER IV, Inc.
+// Copyright 2023-2024 TIER IV, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -22,19 +22,19 @@
 
 namespace behavior_velocity_planner::dynamic_obstacle_stop
 {
-/// @brief find the first collision along an object footprint and the ego path
-/// @param [in] ego_data ego data including its path and footprint
+/// @brief find the collision point closest to ego along an object footprint
+/// @param [in] ego_data ego data including its path and footprints used for finding a collision
 /// @param [in] object_pose pose of the dynamic object
 /// @param [in] object_footprint footprint of the object used for finding a collision
-/// @return the first collision point along the object footprint (if any)
-std::optional<geometry_msgs::msg::Point> find_first_collision_point(
+/// @return the collision point closest to ego (if any)
+std::optional<geometry_msgs::msg::Point> find_closest_collision_point(
   const EgoData & ego_data, const geometry_msgs::msg::Pose & object_pose,
   const tier4_autoware_utils::Polygon2d & object_footprint);
 
-/// @brief find the earliest collision along the ego path and an obstacle footprint
+/// @brief find the earliest collision along the ego path
 /// @param [in] ego_data ego data including its path and footprint
 /// @param [in] objects obstacles
-/// @param [in] obstacle_forward_footprints obstacle footprints
+/// @param [in] obstacle_forward_footprints obstacle footprints to check for collisions
 /// @return the point of earliest collision along the ego path
 std::vector<Collision> find_collisions(
   const EgoData & ego_data,
