+++ planning/behavior_velocity_dynamic_obstacle_stop_module/src/scene_dynamic_obstacle_stop.cpp
@@ -1,4 +1,4 @@
-// Copyright 2023 TIER IV, Inc. All rights reserved.
+// Copyright 2023-2024 TIER IV, Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -27,7 +27,6 @@
 #include <tier4_autoware_utils/geometry/geometry.hpp>
 #include <tier4_autoware_utils/system/stop_watch.hpp>
 
-#include <limits>
 #include <memory>
 #include <string>
 #include <utility>
@@ -45,52 +44,6 @@ DynamicObstacleStopModule::DynamicObstacleStopModule(
 : SceneModuleInterface(module_id, logger, clock), params_(std::move(planner_param))
 {
   velocity_factor_.init(PlanningBehavior::UNKNOWN);
-}
-
-void update_object_map(
-  ObjectStopDecisionMap & object_map, const std::vector<Collision> & collisions,
-  const rclcpp::Time & now,
-  const std::vector<autoware_auto_planning_msgs::msg::PathPointWithLaneId> & path_points,
-  const PlannerParam params)
-{
-  for (auto & [object, decision] : object_map) decision.collision_detected = false;
-  for (const auto & collision : collisions) {
-    if (auto search = object_map.find(collision.object_uuid); search != object_map.end()) {
-      search->second.collision_detected = true;
-      const auto is_closer_collision_point =
-        motion_utils::calcSignedArcLength(
-          path_points, search->second.collision_point, collision.point) < 0.0;
-      if (is_closer_collision_point) search->second.collision_point = collision.point;
-    } else {
-      object_map[collision.object_uuid].collision_point = collision.point;
-    }
-  }
-  for (auto it = object_map.begin(); it != object_map.end();) {
-    auto & decision = it->second;
-    decision.update_timers(now, params.add_duration_buffer, params.remove_duration_buffer);
-    if (decision.is_inactive())
-      it = object_map.erase(it);
-    else
-      ++it;
-  }
-}
-
-std::optional<geometry_msgs::msg::Point> find_earliest_collision(
-  const ObjectStopDecisionMap & object_map, const EgoData & ego_data)
-{
-  std::optional<geometry_msgs::msg::Point> earliest_collision;
-  double earliest_collision_arc_length = std::numeric_limits<double>::max();
-  for (auto & [object_uuid, decision] : object_map) {
-    if (decision.should_be_avoided()) {
-      const auto arc_length = motion_utils::calcSignedArcLength(
-        ego_data.path.points, ego_data.pose.position, decision.collision_point);
-      if (arc_length < earliest_collision_arc_length) {
-        earliest_collision_arc_length = arc_length;
-        earliest_collision = decision.collision_point;
-      }
-    }
-  }
-  return earliest_collision;
 }
 
 bool DynamicObstacleStopModule::modifyPathVelocity(PathWithLaneId * path, StopReason * stop_reason)
