+++ planning/behavior_velocity_dynamic_obstacle_stop_module/src/scene_dynamic_obstacle_stop.hpp
@@ -15,6 +15,7 @@
 #ifndef SCENE_DYNAMIC_OBSTACLE_STOP_HPP_
 #define SCENE_DYNAMIC_OBSTACLE_STOP_HPP_
 
+#include "object_stop_decision.hpp"
 #include "types.hpp"
 
 #include <behavior_velocity_planner_common/scene_module_interface.hpp>
